number 1 fixed

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -12,12 +12,12 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 {
 	if (separator != NULL)
 	{
-		int i, x;
+		unsigned int i, x;
 		va_list args;
 
 		va_start(args, n);
 
-		for (i = 0; i < (int)n; i++)
+		for (i = 0; i < n; i++)
 		{
 			x = va_arg(args, int);
 			printf("%d", x);
